requirements.txt
# Web Server & Real-time Voice
fastapi
uvicorn
fastrtc[vad, stt, tts]
fastrtc-moonshine-onnx
kokoro-onnx

# AI & LLM Logic
ollama
langchain
langchain-ollama
langchain-community
langchain-text-splitters

# Email & Data Processing
imapclient
python-dotenv
requests
pandas
scikit-learn

# Utilities
numpy
python-multipart
